Add random picked color, effects for click events
Picked color is any random color from the given array of colors. 
When the user clicks a square, there is a message display with the info if the guess is correct or not.
When the guess is correct, all colors change to the picked color.
Each click on the incorrect square causes it to change its color to black.